refactor: AttendanceReader 도매인 패키지로 이동
