+++ src/main/java/attendance/Application.java
@@ -3,12 +3,10 @@ package attendance;
 import attendance.controller.AttendanceController;
 import attendance.service.AttendanceService;
 import attendance.service.DateGeneratorImpl;
-import attendance.utils.AttendanceFileParser;
+import attendance.domain.AttendanceFileParser;
 import attendance.utils.ErrorUtils;
 import attendance.view.InputView;
 import attendance.view.OutputView;
-
-import java.time.LocalDate;
 
 public class Application {
