+++ src/test/java/attendance/utils/AttendanceFileParserTest.java
@@ -4,6 +4,8 @@ import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.List;
 
+import attendance.domain.AttendanceFileParser;
+import attendance.domain.AttendanceReader;
 import attendance.dto.FileRequestDto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
